requirements.txt
Click>=8.0
Pillow>=9.0
piexif>=1.1.3